issue #658 move toSigned, cellTypeUnion to GeneralUtils and add test for CellTypeUnionWithNoData
