+++ openeo-geotrellis/src/main/scala/org/openeo/geotrellis/layers/BandCompositeRasterSource.scala
@@ -7,7 +7,7 @@ import geotrellis.raster.{CellType, CropOptions, CroppedTile, GridBounds, GridEx
 import geotrellis.vector.Extent
 import net.jodah.failsafe.event.ExecutionAttemptedEvent
 import net.jodah.failsafe.{Failsafe, RetryPolicy}
-import org.openeo.geotrellis.cellTypeUnion
+import org.openeo.geotrellis.GeneralUtils.cellTypeUnion
 import org.openeo.geotrelliscommon.ResampledTile
 import org.slf4j.LoggerFactory
 import software.amazon.awssdk.core.exception.AbortedException
