+++ openeo-geotrellis/src/main/scala/org/openeo/geotrellis/layers/raster_source/ValueOffsetRasterSource.scala
@@ -4,7 +4,7 @@ import geotrellis.proj4.CRS
 import geotrellis.raster.io.geotiff.OverviewStrategy
 import geotrellis.raster.{CellSize, CellType, GridBounds, GridExtent, MultibandTile, Raster, RasterMetadata, RasterSource, ResampleMethod, ResampleTarget, SourceName, TargetCellType, Tile}
 import geotrellis.vector.Extent
-import org.openeo.geotrellis.toSigned
+import org.openeo.geotrellis.GeneralUtils.toSigned
 import org.slf4j.LoggerFactory
 
 /**
